requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
playwright>=1.40.0
asyncio-mqtt>=0.13.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
strands-agents
strands-agents-tools
pyyaml>=6.0